requirements.txt
flask
openai
python-docx
PyMuPDF
